Create README.md
Data preparation scripts were developed to interpolate microclim's original resolution from 15 km to 1 and 5km. The purpose is to find out if different resolutions affect the detection of signals that could be lost at coarser resolutions.

We begin to explore this phenomenon with the exploratory analysis in which we use land use maps at the original resolution of 30m and decrease it up to 5km.

The results of these analyses will be then used to decide which is the optimal resolution to model snake distributions and snakebite risk.